clock finalization
