+++ .gitignore
@@ -1,1 +1,2 @@
 .vscode
+IT_Academy_project_hw
