+++ requiremets.txt
@@ -1,0 +1,1 @@
+colorama==0.4.6
